requirements.txt
fastapi
uvicorn
streamlit
pydantic
pandas
python-dotenv
google-cloud-storage
anthropic
